+++ group3_genetic/src/main/java/fr/isen/cir56/group3_genetic/Wizard/Configurator/ParameterPanel/ParameterProbabilityChangedListener.java
@@ -15,7 +15,7 @@ public class ParameterProbabilityChangedListener extends JPanelChangedListener<P
 	
 	@Override
 	public void valueChanged(int oldValue, int newValue) {
-		this.controller.setValue((double) newValue);
+		this.controller.setValue(newValue);
 	}
 	
 }
